# FanApp/FanApp/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FanClubAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Setlists": {
      "get": {
        "tags": [
          "Setlists"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Setlists"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Setlists"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Setlists"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Setlists"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Setlists"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Setlists"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Setlists"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Setlists"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Setlists"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Setlists"
                }
              }
            }
          }
        }
      }
    },
    "/api/Setlists/{id}": {
      "get": {
        "tags": [
          "Setlists"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Setlists"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Setlists"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Setlists"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Setlists"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Setlists"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Setlists"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Setlists"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Setlists"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Songs": {
      "get": {
        "tags": [
          "Songs"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Songs"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Songs"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Songs"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Songs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Songs"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Songs"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Songs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Songs"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Songs"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Songs"
                }
              }
            }
          }
        }
      }
    },
    "/api/Songs/{id}": {
      "get": {
        "tags": [
          "Songs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Songs"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Songs"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Songs"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Songs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Songs"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Songs"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Songs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Songs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Videos": {
      "get": {
        "tags": [
          "Videos"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Videos"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Videos"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Videos"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Videos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Videos"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Videos"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Videos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Videos"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Videos"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Videos"
                }
              }
            }
          }
        }
      }
    },
    "/api/Videos/{id}": {
      "get": {
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Videos"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Videos"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Videos"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Videos"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Videos"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Videos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Albums": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Songs"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Concerts": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "concertDate": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "venue": {
            "type": "string",
            "nullable": true
          },
          "ticketUrl": {
            "type": "string",
            "nullable": true
          },
          "setlists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Setlists"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DateOnly": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "dayOfYear": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DayOfWeek": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Setlists": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "concertId": {
            "type": "integer",
            "format": "int32"
          },
          "songId": {
            "type": "integer",
            "format": "int32"
          },
          "songOrder": {
            "type": "integer",
            "format": "int32"
          },
          "concert": {
            "$ref": "#/components/schemas/Concerts"
          },
          "song": {
            "$ref": "#/components/schemas/Songs"
          }
        },
        "additionalProperties": false
      },
      "Songs": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "albumId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "nullable": true
          },
          "trackNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lyrics": {
            "type": "string",
            "nullable": true
          },
          "album": {
            "$ref": "#/components/schemas/Albums"
          },
          "setlists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Setlists"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Videos": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "youtubeUrl": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}